# backend/openapi.yaml
components:
  schemas:
...
    controller.apiResponse-any:
      properties:
        data: {}
        error:
          type: string
      type: object
    controller.apiResponse-array_int32:
      properties:
        data:
          items:
            type: integer
          type: array
          uniqueItems: false
        error:
          type: string
      type: object
...
    controller.apiResponse-array_model_User:
      properties:
        data:
          items:
            $ref: '#/components/schemas/model.User'
          type: array
          uniqueItems: false
        error:
          type: string
      type: object
...
    controller.apiResponse-model_User:
      properties:
        data:
          $ref: '#/components/schemas/model.User'
        error:
          type: string
      type: object
    controller.apiResponse-model_UserFismaSystem:
      properties:
        data:
          $ref: '#/components/schemas/model.UserFismaSystem'
        error:
          type: string
      type: object
...
    model.User:
      properties:
        assignedopdivids:
          items:
            type: integer
          type: array
          uniqueItems: false
        deleted:
          type: boolean
        email:
          type: string
        fullname:
          type: string
        identity_provider:
          type: string
        role:
          type: string
        userid:
          type: string
      type: object
    model.UserFismaSystem:
      properties:
        fismasystemid:
          type: integer
        userid:
          type: string
      type: object
...
  securitySchemes:
    bearerAuth:
      in: header
      name: Authorization
      type: apiKey
...
info:
  description: API for the Zero Trust Maturity Framework (ZTMF).
  title: ZTMF API
  version: 1.0.0
openapi: 3.1.0
paths:
...
  /users:
    get:
      parameters:
      - description: Filter by email (partial match)
        in: query
        name: email
        schema:
          type: string
      - description: Filter by full name (partial match)
        in: query
        name: fullname
        schema:
          type: string
      - description: Filter by role
        in: query
        name: role
        schema:
          type: string
      - description: Include soft-deleted users
        in: query
        name: deleted
        schema:
          type: boolean
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/controller.apiResponse-array_model_User'
          description: OK
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/controller.apiResponse-any'
          description: Forbidden
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/controller.apiResponse-any'
          description: Internal Server Error
      security:
      - bearerAuth: []
      summary: List all users
      tags:
      - users
    post:
      parameters:
      - description: User ID (for update)
        in: path
        name: userid
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/model.User'
                description: User to create or update
                summary: body
        description: User to create or update
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/controller.apiResponse-model_User'
          description: Created
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/controller.apiResponse-any'
          description: Bad Request
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/controller.apiResponse-any'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/controller.apiResponse-any'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/controller.apiResponse-any'
          description: Internal Server Error
      security:
      - bearerAuth: []
      summary: Create or update a user
      tags:
      - users
  /users/{userid}:
    delete:
      parameters:
      - description: User ID
        in: path
        name: userid
        required: true
        schema:
          type: string
      responses:
        "204":
          description: No Content
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/controller.apiResponse-any'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/controller.apiResponse-any'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/controller.apiResponse-any'
          description: Internal Server Error
      security:
      - bearerAuth: []
      summary: Delete a user
      tags:
      - users
    get:
      parameters:
      - description: User ID
        in: path
        name: userid
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/controller.apiResponse-model_User'
          description: OK
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/controller.apiResponse-any'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/controller.apiResponse-any'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/controller.apiResponse-any'
          description: Internal Server Error
      security:
      - bearerAuth: []
      summary: Get a user by ID
      tags:
      - users
    put:
      parameters:
      - description: User ID (for update)
        in: path
        name: userid
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/model.User'
                description: User to create or update
                summary: body
        description: User to create or update
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/controller.apiResponse-model_User'
          description: Created
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/controller.apiResponse-any'
          description: Bad Request
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/controller.apiResponse-any'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/controller.apiResponse-any'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/controller.apiResponse-any'
          description: Internal Server Error
      security:
      - bearerAuth: []
      summary: Create or update a user
      tags:
      - users
  /users/{userid}/assignedfismasystems:
    get:
      parameters:
      - description: User ID
        in: path
        name: userid
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/controller.apiResponse-array_int32'
          description: OK
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/controller.apiResponse-any'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/controller.apiResponse-any'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/controller.apiResponse-any'
          description: Internal Server Error
      security:
      - bearerAuth: []
      summary: List the FISMA system ids assigned to a user
      tags:
      - users
    post:
      parameters:
      - description: User ID
        in: path
        name: userid
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/model.UserFismaSystem'
                description: FISMA system assignment
                summary: body
        description: FISMA system assignment
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/controller.apiResponse-model_UserFismaSystem'
          description: Created
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/controller.apiResponse-any'
          description: Bad Request
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/controller.apiResponse-any'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/controller.apiResponse-any'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/controller.apiResponse-any'
          description: Internal Server Error
      security:
      - bearerAuth: []
      summary: Assign a FISMA system to a user
      tags:
      - users
  /users/{userid}/assignedfismasystems/{fismasystemid}:
    delete:
      parameters:
      - description: User ID
        in: path
        name: userid
        required: true
        schema:
          type: string
      - description: FISMA System ID
        in: path
        name: fismasystemid
        required: true
        schema:
          type: integer
      responses:
        "204":
          description: No Content
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/controller.apiResponse-any'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/controller.apiResponse-any'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/controller.apiResponse-any'
          description: Internal Server Error
      security:
      - bearerAuth: []
      summary: Unassign a FISMA system from a user
      tags:
      - users
...
  /users/{userid}/restore:
    put:
      parameters:
      - description: User ID
        in: path
        name: userid
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/controller.apiResponse-model_User'
          description: OK
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/controller.apiResponse-any'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/controller.apiResponse-any'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/controller.apiResponse-any'
          description: Internal Server Error
      security:
      - bearerAuth: []
      summary: Restore a soft-deleted user
      tags:
      - users
  /users/current:
    get:
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/controller.apiResponse-model_User'
          description: OK
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/controller.apiResponse-any'
          description: Internal Server Error
      security:
      - bearerAuth: []
      summary: Get the currently authenticated user
      tags:
      - users
servers:
- url: /api/v1